!IMPORTANT everything fixed this time for realz. Maybe.
Added a new function for creating savegames as the previous would have
transformed into a mess (and it didn't work) updated stuff to reflect
the new menubuttons and added a "quit" function. Important that. The
levels in that folder is now moved in a way that actually works. That
should be an advantage, right?